style: Apply advanced code-splitting techniques
Well, they're not that special.

Removes HomeScreen from App.tsx
